+++ Runtime/Renderer.h
@@ -45,7 +45,7 @@ public:
 		return (UINT)_effects.size();
 	}
 
-	const EffectDesc& GetEffectDesc(size_t idx) const noexcept;
+	const EffectDesc& GetEffectDesc(UINT idx) const noexcept;
 
 private:
 	bool _InitializeOverlayDrawer();
